refactor: extract tui app state
